requirements.txt
addict==2.4.0
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
click==8.1.3
comm==0.1.2
ConfigArgParse==1.5.3
contourpy==1.0.6
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
fonttools==4.38.0
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mpmath==1.2.1
nbformat==5.5.0
nest-asyncio==1.5.6
numpy==1.24.1
open3d==0.16.1
opencv-python==4.7.0.68
packaging==22.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
plotly==5.12.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
scikit-learn==1.2.1
scipy==1.9.3
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tenacity==8.1.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
wcwidth==0.2.5
Werkzeug==2.2.2
